nonii nyt siirretty tonne samaan
