+++ src/components/ParameterTabsComponent.js
@@ -16,7 +16,6 @@ import WayareaPolygonContext from "contexts/WayareaPolygonContext";
 import parametersValidationSchema from "constants/ParametersValidationSchema";
 import SelectedCalculationTypeContext from "contexts/SelectedCalculationTypeContext";
 import { calculationTypeEnums } from "constants/enums";
-import { GeneratePaths } from "./GeneratePathsComponent";
 
 function a11yProps(index) {
   return {
@@ -40,9 +39,52 @@ export default function ParameterTabsComponent() {
   const { selectedCalculationType } = useContext(
     SelectedCalculationTypeContext
   );
+  const generatePaths = async (
+    values,
+    selectedCalculationType,
+    selectedRouteline
+  ) => {
+    let paths = {
+      path: "",
+      path_wayarea: "",
+      path_navigationline: "",
+      path_wayarea_navigationline: "",
+      path_routeline: "",
+      path_wayarea_routeline: "",
+    };
 
+    if (selectedCalculationType == calculationTypeEnums.ROUTELINE) {
+      paths.path = `routeline/calculate_risk?routename=${encodeURIComponent(
+        selectedRouteline
+      )}`;
+      paths.path_wayarea = `routeline/wayarea_polygons?routename=${encodeURIComponent(
+        selectedRouteline
+      )}`;
+    } else if (selectedCalculationType == calculationTypeEnums.NAVIGATIONLINE) {
+      paths.path = `fairway/calculate_risk?vaylat=${encodeURIComponent(
+        values.vaylat
+      )}`;
+      paths.path_wayarea = `wayarea?vaylat=${encodeURIComponent(
+        values.vaylat
+      )}`;
+    } else if (selectedCalculationType == calculationTypeEnums.COMPARE) {
+      paths.path_navigationline = `fairway/calculate_risk?vaylat=${encodeURIComponent(
+        values.vaylat
+      )}`;
+      paths.path_wayarea_navigationline = `wayarea?vaylat=${encodeURIComponent(
+        values.vaylat
+      )}`;
+      paths.path_routeline = `routeline/calculate_risk?routename=${encodeURIComponent(
+        selectedRouteline
+      )}`;
+      paths.path_wayarea_routeline = `routeline/wayarea_polygons?routename=${encodeURIComponent(
+        selectedRouteline
+      )}`;
+    }
+    return paths;
+  };
   const fetchRiskValue = async (values) => {
-    const paths = await GeneratePaths(
+    const paths = await generatePaths(
       values,
       selectedCalculationType,
       selectedRouteline
